requirements.txt
antlr4-python3-runtime==4.8
atomicwrites @ file:///home/conda/feedstock_root/build_artifacts/atomicwrites_1588182545583/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620323180557/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
certifi==2020.12.5
click==7.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
coverage @ file:///D:/bld/coverage_1614551455218/work
cycler==0.10.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1618380339289/work
hydra-core==1.0.6
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1603384189793/work
ipykernel @ file:///D:/bld/ipykernel_1620254623900/work/dist/ipykernel-5.5.4-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1620254727263/work
ipython-genutils==0.2.0
jedi @ file:///D:/bld/jedi_1610147001629/work
joblib==1.0.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1615693636836/work
jupyter-core @ file:///D:/bld/jupyter_core_1612125412016/work
kiwisolver @ file:///D:/bld/kiwisolver_1610099953470/work
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
matplotlib @ file:///D:/bld/matplotlib-suite_1617212922200/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
# Editable install with no version control (ml-project==0.1.0)
-e c:\storage\san\scripts\anacondaprojects\learning\made\semester2\ml_in_prod\cc\ml_project
more-itertools @ file:///home/conda/feedstock_root/build_artifacts/more-itertools_1619297856442/work
mypy-extensions==0.4.3
numpy @ file:///D:/bld/numpy_1616933442622/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.0.6
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandas @ file:///D:/bld/pandas_1618263878420/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
patsy==0.5.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///D:/bld/pillow_1617818313335/work
pluggy @ file:///D:/bld/pluggy_1610318302422/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1616432837031/work
py @ file:///home/conda/feedstock_root/build_artifacts/py_1607783655754/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1620245170812/work
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pytest==6.2.4
pytest-cov @ file:///home/conda/feedstock_root/build_artifacts/pytest-cov_1611167431813/work
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
pywin32==300
PyYAML==5.4.1
pyzmq @ file:///D:/bld/pyzmq_1614611981521/work
scikit-learn==0.24.2
scipy @ file:///C:/bld/scipy_1619562066759/work
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1611834504644/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
# Editable install with no version control (src==0.1.0)
-e c:\storage\san\scripts\anacondaprojects\learning\made\semester2\ml_in_prod\cc\ml_project
statsmodels @ file:///D:/bld/statsmodels_1612273747431/work
threadpoolctl==2.1.0
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tornado @ file:///D:/bld/tornado_1610094910414/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
typing-extensions==3.10.0.0
typing-inspect==0.6.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
wincertstore==0.2